+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   + Expose http module. New method available : `lunaris.http.request`.
   + Add attribute sourceWhere for filters in order to filter the filter's data.
   + Add index of map id in each collection. It allows to not duplicate values which have the same map id. If duplication, the duplicate value is merged with the previous one.
+  + Directive v-lunaris supports ckeckbox input.
 - Fix:
   + The redirection could fail.
   + Add parameter isUnique for `lunaris.hook(hook, handkler, isUnique)`. If true, the function will register unique handlers.